src: gpu: bnorm: gen9: use char dt for relu workspace

+++ src/gpu/gpu_batch_normalization_pd.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2019-2020 Intel Corporation
+* Copyright 2019-2021 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -33,7 +33,7 @@ template <typename pd_t>
 inline void gpu_init_default_ws(pd_t *self, memory_desc_t &ws_md) {
     auto mdw = memory_desc_wrapper(self->src_md(0));
     ws_md = *mdw.md_;
-    ws_md.data_type = data_type::s32;
+    ws_md.data_type = data_type::s8;
 }
 } // namespace
